+++ demisto_sdk/commands/common/hook_validations/integration.py
@@ -3,7 +3,6 @@ import re
 from typing import Dict, Optional
 
 import yaml
-
 from demisto_sdk.commands.common.constants import (
     BANG_COMMAND_ARGS_MAPPING_DICT, BANG_COMMAND_NAMES, DBOT_SCORES_DICT,
     DEPRECATED_REGEXES, ENDPOINT_COMMAND_NAME, ENDPOINT_FLEXIBLE_REQUIRED_ARGS,
@@ -962,7 +961,8 @@ class IntegrationValidator(ContentEntityValidator):
             (bool).
         """
         display_name = self.current_file.get('display')
-        return all([self.is_valid_versioned_display_name(), not self.name_contains_contributor_type_name(display_name)])
+        return all([self.is_valid_versioned_display_name(),
+                    self.name_does_not_contain_contributor_type_name(display_name)])
 
     def is_valid_versioned_display_name(self) -> bool:
         version_number: Optional[str] = get_file_version_suffix_if_exists(self.current_file,