+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Added to **validate** command validation of integration/script/pack name not containing contributor type name
 
 # 1.4.4
 * When formatting incident types with Auto-Extract rules and without mode field, the **format** command will now add the user selected mode.